requirements.txt
tensorflow 
pillow